+++ lib/versionone_sdk/parser_xml_assets.rb
@@ -63,6 +63,9 @@ module VersiononeSdk
             oAsset.setProp(yPropKey,xxPropVal)
           end
         else
+          if oNodeChild.name == 'Message' && oNodeChild.text == 'Not Found'
+            raise RuntimeError, "E_ASSET_NOT_FOUND #{oNodeAsset.attribute('href').value}"
+          end
           raise RuntimeError, "E_UNKNOWN_ASSET_NODE_NAME #{oNodeChild.name}"
         end
       end